Задание Jdbc. Учел все замечания.
