+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/model/Genre.java
@@ -15,12 +15,9 @@ public class Genre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
     private long id;
 
-    @Column(name = "name")
     private String name;
 
-    @Column(name = "BOOK_ID")
-    private long bookId;
+    private long book_Id;
 }
