+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/shell/IAuthors.java
@@ -1,0 +1,16 @@
+package ru.otus.spring.shell;
+
+import ru.otus.spring.model.Authors;
+import java.util.List;
+
+public interface IAuthors {
+
+    public void getAutors();
+
+    public void getAutorsCount();
+
+    public void getAutorsGetId();
+
+    public void aPrint(List<Authors> authors);
+
+}
