+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/shell/IBooks.java
@@ -1,0 +1,20 @@
+package ru.otus.spring.shell;
+
+import ru.otus.spring.model.Books;
+import java.util.List;
+
+public interface IBooks {
+    public void count();
+
+    public void getAll();
+
+    public void getBookById();
+
+    public void delBook();
+
+    public void updateBook();
+
+    public void insertBook();
+
+    public void bPrint(List<Books> books);
+}
