+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/shell/IGenre.java
@@ -1,0 +1,17 @@
+package ru.otus.spring.shell;
+
+import ru.otus.spring.model.Genre;
+import java.util.List;
+
+public interface IGenre {
+
+    public void getGenre();
+
+    public void getGenreCount();
+
+    public void getGenreGetId();
+
+    public void gPrint(List<Genre> genres);
+
+
+}
